Add model for supplier assessment appointment
This extracts most of the action plan appointment model into an
Appointment interface.
