+++ server/models/appointment.ts
@@ -1,0 +1,12 @@
+import AppointmentAttendance from './appointmentAttendance'
+import AppointmentBehaviour from './appointmentBehaviour'
+
+export default interface Appointment {
+  appointmentTime: string | null
+  durationInMinutes: number | null
+  sessionFeedback: {
+    attendance: AppointmentAttendance
+    behaviour: AppointmentBehaviour
+    submitted: boolean
+  }
+}
