+++ server/models/appointmentAttendance.ts
@@ -1,0 +1,6 @@
+export type Attended = 'yes' | 'no' | 'late' | null
+
+export default interface AppointmentAttendance {
+  attended: Attended
+  additionalAttendanceInformation: string | null
+}
