+++ server/models/appointmentBehaviour.ts
@@ -1,0 +1,4 @@
+export default interface AppointmentBehaviour {
+  behaviourDescription: string | null
+  notifyProbationPractitioner: boolean | null
+}
