+++ server/services/interventionsService.ts
@@ -10,7 +10,9 @@ import Intervention from '../models/intervention'
 import PCCRegion from '../models/pccRegion'
 import CancellationReason from '../models/cancellationReason'
 import ServiceCategory from '../models/serviceCategory'
-import ActionPlan, { ActionPlanAppointment, AppointmentAttendance, AppointmentBehaviour } from '../models/actionPlan'
+import ActionPlan, { ActionPlanAppointment } from '../models/actionPlan'
+import AppointmentAttendance from '../models/appointmentAttendance'
+import AppointmentBehaviour from '../models/appointmentBehaviour'
 import DraftReferral from '../models/draftReferral'
 import SentReferral from '../models/sentReferral'
 import ReferralDesiredOutcomes from '../models/referralDesiredOutcomes'
